fix: rename datasource to database
