+++ src/database/database.module.ts
@@ -1,9 +1,13 @@
-import { Global, Module } from '@nestjs/common';
+import { forwardRef, Module } from '@nestjs/common';
+import { DatabaseController } from './database.controller';
 import { DatabaseService } from './database.service';
+import { DatabaseGateway } from './database.gateway';
+import { TemplateModule } from 'src/template/template.module';
 
-@Global()
 @Module({
-  providers: [DatabaseService],
+  imports: [forwardRef(() => TemplateModule)],
+  controllers: [DatabaseController],
+  providers: [DatabaseService, DatabaseGateway],
   exports: [DatabaseService],
 })
 export class DatabaseModule {}
