--- src/datasource/dto/index.ts
@@ -1,1 +1,0 @@
-export * from './datasource.dto';
